refactor: extract elevated-spawn helper into a shared module
No functional change -- pulls spawnElevatedOnWindows() and its supporting
functions out of launch-game.ts into src/main/helpers/spawn-elevated-windows.ts
so the same elevation-retry logic can be reused by the installer launch path
(open-game-installer.ts), which needs it just as much as game launches do.
